Add additional CEC settings
This PR exposes some more knobs to twiddle when talking to `cec-client`,
namely, the CEC services will now set the "On Screen Display" name to
match the system's hostname (`bazzite` by default).

And because _my_ particular PulseEight HDMI CEC Adaptor is fussy, I need
to override the HDMI port that that it uses for determining the active
source. I've also exposed this as an option which can be set.

Lastly, because the arguments list to `cec-client` can be osbcure, and
quite long, I've used a Bash array to store the long version of the arg
names to make the logic easier to follow, and splatting the args into
calls to `cec-client` so that changing these arguments in the future is
much easier.